#18236 refactoriing to extract code needed to perforn import from upgrade task
