+++ dotCMS/src/main/java/com/dotcms/security/apps/AppsAPIImpl.java
@@ -1,7 +1,13 @@
 package com.dotcms.security.apps;
 
+import static com.dotcms.security.apps.AppDescriptorHelper.getUserAppsDescriptorDirectory;
+import static com.dotcms.security.apps.AppsUtil.exportSecret;
+import static com.dotcms.security.apps.AppsUtil.importSecrets;
+import static com.dotcms.security.apps.AppsUtil.internalKey;
+import static com.dotcms.security.apps.AppsUtil.mapForValidation;
 import static com.dotcms.security.apps.AppsUtil.readJson;
 import static com.dotcms.security.apps.AppsUtil.toJsonAsChars;
+import static com.dotcms.security.apps.AppsUtil.validateForSave;
 import static com.dotmarketing.util.UtilMethods.isNotSet;
 import static com.dotmarketing.util.UtilMethods.isSet;
 import static com.google.common.collect.ImmutableList.of;
@@ -17,7 +23,6 @@ import com.dotmarketing.common.model.ContentletSearch;
 import com.dotmarketing.exception.AlreadyExistException;
 import com.dotmarketing.exception.DoesNotExistException;
 import com.dotmarketing.exception.DotDataException;
-import com.dotmarketing.exception.DotDataValidationException;
 import com.dotmarketing.exception.DotRuntimeException;
 import com.dotmarketing.exception.DotSecurityException;
 import com.dotmarketing.exception.InvalidLicenseException;
@@ -25,39 +30,22 @@ import com.dotmarketing.portlets.contentlet.business.ContentletAPI;
 import com.dotmarketing.portlets.contentlet.business.HostAPI;
 import com.dotmarketing.util.Config;
 import com.dotmarketing.util.Logger;
-import com.dotmarketing.util.UtilMethods;
-import com.fasterxml.jackson.core.JsonParser.Feature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMap.Builder;
-import com.google.common.collect.Sets;
-import com.google.common.collect.Sets.SetView;
 import com.liferay.portal.model.User;
-import com.liferay.util.EncryptorException;
 import com.liferay.util.StringPool;
-import io.vavr.Tuple;
 import io.vavr.Tuple2;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.io.OutputStream;
 import java.net.URISyntaxException;
-import java.net.URL;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.Key;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -65,7 +53,6 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -83,55 +70,35 @@ public class AppsAPIImpl implements AppsAPI {
     private final LocalSystemEventsAPI localSystemEventsAPI;
 
     private final LicenseValiditySupplier licenseValiditySupplier;
-
-    private final ObjectMapper ymlMapper = new ObjectMapper(new YAMLFactory())
-            .enable(Feature.STRICT_DUPLICATE_DETECTION)
-            //.enable(SerializationFeature.INDENT_OUTPUT)
-            .findAndRegisterModules();
+    private final AppDescriptorHelper appDescriptorHelper;
 
     @VisibleForTesting
-    public AppsAPIImpl( final LayoutAPI layoutAPI, final HostAPI hostAPI, final ContentletAPI contentletAPI,
-            final SecretsStore secretsRepository, final AppsCache appsCache, final LocalSystemEventsAPI localSystemEventsAPI, final LicenseValiditySupplier licenseValiditySupplier) {
+    public AppsAPIImpl(final LayoutAPI layoutAPI, final HostAPI hostAPI,
+            final ContentletAPI contentletAPI,
+            final SecretsStore secretsRepository, final AppsCache appsCache,
+            final LocalSystemEventsAPI localSystemEventsAPI,
+            final AppDescriptorHelper appDescriptorHelper,
+            final LicenseValiditySupplier licenseValiditySupplier) {
         this.layoutAPI = layoutAPI;
         this.hostAPI = hostAPI;
         this.contentletAPI = contentletAPI;
         this.secretsStore = secretsRepository;
         this.appsCache = appsCache;
         this.localSystemEventsAPI = localSystemEventsAPI;
+        this.appDescriptorHelper = appDescriptorHelper;
         this.licenseValiditySupplier = licenseValiditySupplier;
     }
 
+    /**
+     * default constructor
+     */
     public AppsAPIImpl() {
         this(APILocator.getLayoutAPI(), APILocator.getHostAPI(),
                 APILocator.getContentletAPI(), SecretsStore.INSTANCE.get(),
                 CacheLocator.getAppsCache(), APILocator.getLocalSystemEventsAPI(),
+                new AppDescriptorHelper(),
                 new LicenseValiditySupplier() {
                 });
-    }
-
-    /**
-     * One single method takes care of building the internal-key
-     */
-    private String internalKey(final String serviceKey, final Host host) {
-        return internalKey(serviceKey, host == null ? null : host.getIdentifier());
-    }
-
-    /**
-     * Given a service key and an identifier this builds an internal key composed by the two values concatenated
-     * And lowercased.
-     * Like `5e096068-edce-4a7d-afb1-95f30a4fa80e:serviceKeyNameXYZ`
-     * @param serviceKey
-     * @param hostIdentifier
-     * @return
-     */
-    private String internalKey(final String serviceKey, final String hostIdentifier) {
-        // if Empty ServiceKey is passed everything will be set under systemHostIdentifier:dotCMSGlobalService
-        //Otherwise the internal Key will look like:
-        // `5e096068-edce-4a7d-afb1-95f30a4fa80e:serviceKeyNameXYZ` where the first portion is the hostId
-        final String key = isSet(serviceKey) ? serviceKey : DOT_GLOBAL_SERVICE;
-        final String identifier =
-                (null == hostIdentifier) ? APILocator.systemHost().getIdentifier() : hostIdentifier;
-        return (identifier + HOST_SECRET_KEY_SEPARATOR + key).toLowerCase();
     }
 
     private boolean userDoesNotHaveAccess(final User user) throws DotDataException {
@@ -199,13 +166,18 @@ public class AppsAPIImpl implements AppsAPI {
         return stream.collect(Collectors.groupingBy(strings -> strings[0],
                 Collectors.mapping(strings -> strings[1], Collectors.toSet())));
     }
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Optional<AppSecrets> getSecrets(final String key,
             final Host host, final User user) throws DotDataException, DotSecurityException {
             return getSecrets(key, false, host, user);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Optional<AppSecrets> getSecrets(final String key,
             final boolean fallbackOnSystemHost,
@@ -301,6 +273,9 @@ public class AppsAPIImpl implements AppsAPI {
         }
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public void saveSecret(final String key, final Tuple2<String, Secret> keyAndSecret,
             final Host host, final User user)
@@ -425,7 +400,7 @@ public class AppsAPIImpl implements AppsAPI {
         synchronized (AppsAPIImpl.class) {
             return appsCache.getAppDescriptorsMeta(() -> {
                 try {
-                    return loadAppDescriptors();
+                    return appDescriptorHelper.loadAppDescriptors();
                 } catch (IOException | URISyntaxException e) {
                     Logger.error(AppsAPIImpl.class,
                             "An error occurred while loading the service descriptor yml files. ",
@@ -481,9 +456,9 @@ public class AppsAPIImpl implements AppsAPI {
         }
         Logger.debug(AppsAPIImpl.class, () -> " ymlFiles are set under:  " + ymlFilesPath);
 
-            final AppSchema appSchema = readAppFile(file.toPath());
+            final AppSchema appSchema = appDescriptorHelper.readAppFile(file.toPath());
             // Now validate the incoming file.. see if we're rewriting an existing file or attempting to re-use an already in use service-key.
-            if (validateAppDescriptor(appSchema)) {
+            if (appDescriptorHelper.validateAppDescriptor(appSchema)) {
                 final File incomingFile = new File(basePath, file.getName());
                 if (incomingFile.exists()) {
                     throw new AlreadyExistException(
@@ -492,43 +467,12 @@ public class AppsAPIImpl implements AppsAPI {
                                     incomingFile.getName()));
                 }
 
-                writeAppFile(incomingFile, appSchema);
+                appDescriptorHelper.writeAppFile(incomingFile, appSchema);
 
                 invalidateCache();
             }
             return new AppDescriptorImpl(file.getName(), false, appSchema);
 
-    }
-
-    /**
-     * There's a version of the method readValue on the ymlMapper which takes a file and internally creates directly a FileInputStream
-     * According to https://dzone.com/articles/fileinputstream-fileoutputstream-considered-harmful
-     * that's very harmful
-     * @param file
-     * @return
-     * @throws DotDataException
-     */
-    private AppSchema readAppFile(final Path file) throws DotDataException {
-        try (InputStream inputStream = Files.newInputStream(file)) {
-            return ymlMapper.readValue(inputStream, AppSchema.class);
-        }catch (Exception e){
-            throw new DotDataException(e.getMessage(), e);
-        }
-    }
-    /**
-     * There's a version of the method writeValue on the ymlMapper which takes a file and internally creates directly a FileOutputStream
-     * According to https://dzone.com/articles/fileinputstream-fileoutputstream-considered-harmful
-     * that's very harmful
-     * @param file
-     * @return
-     * @throws DotDataException
-     */
-    private void writeAppFile(final File file, final AppSchema appSchema) throws DotDataException {
-        try (OutputStream outputStream = Files.newOutputStream(Paths.get(file.getPath()))) {
-             ymlMapper.writeValue(outputStream, appSchema);
-        }catch (Exception e){
-            throw new DotDataException(e.getMessage(), e);
-        }
     }
 
 
@@ -698,302 +642,6 @@ public class AppsAPIImpl implements AppsAPI {
     }
 
     /**
-     * returns DotCMS server folder
-     * @return
-     */
-    public static Path getServerDirectory() {
-        return Paths.get(APILocator.getFileAssetAPI().getRealAssetsRootPath()
-                + File.separator + SERVER_DIR_NAME + File.separator).normalize();
-    }
-
-    /**
-     * This is the directory intended for customers use
-     * @return
-     */
-    private static Path getAppsDefaultDirectory() {
-        return Paths.get(getServerDirectory() + File.separator + APPS_DIR_NAME + File.separator).normalize();
-    }
-
-    /**
-     * This is the directory intended for customers use.
-     * with the option to read an override property from the config
-     * @return
-     */
-    private static Path getUserAppsDescriptorDirectory() {
-        final Supplier<String> supplier = ()-> getAppsDefaultDirectory().toString();
-        final String dirPath = Config
-                .getStringProperty(APPS_DIR_PATH_KEY, supplier.get());
-        return Paths.get(dirPath).normalize();
-    }
-
-    /**
-     * This is the Apps-System-Folder which is meant to hold system apps.
-     * Those that can not be override and are always available.
-     * @return
-     */
-    static Path getSystemAppsDescriptorDirectory() throws URISyntaxException, IOException {
-        final URL res = Thread.currentThread().getContextClassLoader().getResource("apps");
-        if(res == null) {
-            throw new IOException("Unable to find Apps System folder. It should be at /WEB-INF/classes/apps ");
-        } else {
-            return Paths.get(res.toURI()).toAbsolutePath();
-        }
-    }
-
-    /**
-     *  This will get you a list with all the available app-yml files registered in the system.
-     *
-     * @return
-     * @throws IOException
-     * @throws URISyntaxException
-     */
-    private Set<Tuple2<Path, Boolean>> listAvailableYamlFiles() throws IOException, URISyntaxException {
-        final Path systemAppsDescriptorDirectory = getSystemAppsDescriptorDirectory();
-        final Set<Path> systemFiles = listFiles(systemAppsDescriptorDirectory);
-
-        final Path appsDescriptorDirectory = getUserAppsDescriptorDirectory();
-        final File basePath = appsDescriptorDirectory.toFile();
-        if (!basePath.exists()) {
-            basePath.mkdirs();
-        }
-        Logger.debug(AppsAPIImpl.class,
-                () -> " ymlFiles are set under:  " + basePath.toString());
-        final Set<Path> userFiles = listFiles(appsDescriptorDirectory);
-
-        final Set<Path> systemFileNames = systemFiles.stream().map(Path::getFileName)
-                .collect(Collectors.toSet());
-        final Set<Path> filteredUserFiles = userFiles.stream()
-                .filter(path -> systemFileNames.stream().noneMatch(
-                        systemPath -> systemPath.toString()
-                                .equalsIgnoreCase((path.getFileName().toString().toLowerCase()))))
-                .collect(Collectors.toSet());
-
-        return Stream.concat(systemFiles.stream().map(path -> Tuple.of(path, true)),
-                filteredUserFiles.stream().map(path -> Tuple.of(path, false)))
-                .collect(Collectors.toSet());
-    }
-
-    private static DirectoryStream.Filter<Path> filter = new DirectoryStream.Filter<Path>() {
-
-        private static final String ignorePrefix = "_ignore_";
-        private static final String yml = "yml";
-        private static final String yaml = "yaml";
-
-        @Override
-        public boolean accept(final Path path) {
-            if (Files.isDirectory(path)) {
-              return false;
-            }
-            final String fileName = path.getFileName().toString();
-            return !fileName.startsWith(ignorePrefix) && (fileName.endsWith(yaml) || fileName.endsWith(yml)) ;
-        }
-    };
-
-    private Set<Path> listFiles(final Path dir) throws IOException {
-        final Set<Path> fileList = new HashSet<>();
-        try (DirectoryStream<Path> stream = Files.newDirectoryStream(dir, filter)) {
-            stream.forEach(fileList::add);
-        }
-        return fileList;
-    }
-
-    private List<AppDescriptor> loadAppDescriptors()
-            throws IOException, URISyntaxException {
-
-        final ImmutableList.Builder<AppDescriptor> builder = new ImmutableList.Builder<>();
-        final Set<Tuple2<Path, Boolean>> filePaths = listAvailableYamlFiles();
-        for (final Tuple2<Path, Boolean> filePath : filePaths) {
-            try {
-                final Path path = filePath._1;
-                final boolean systemApp = filePath._2;
-                final AppSchema appSchema = readAppFile(path);
-                if (validateAppDescriptor(appSchema)) {
-                    builder.add(new AppDescriptorImpl(path.getFileName().toString(), systemApp, appSchema));
-                }
-            } catch (Exception e) {
-                Logger.error(AppsAPIImpl.class,
-                        String.format("Error reading yml file `%s`.", filePath), e);
-            }
-        }
-
-        return builder.build();
-    }
-
-    /**
-     * internal descriptor validator
-     * @param appDescriptor
-     * @return
-     * @throws DotDataValidationException
-     */
-   private boolean validateAppDescriptor(final AppSchema appDescriptor)
-           throws DotDataValidationException {
-
-       final List<String> errors = new ArrayList<>();
-
-       if(isNotSet(appDescriptor.getName())){
-           errors.add("The required field `name` isn't set on the incoming file.");
-       }
-
-       if(isNotSet(appDescriptor.getDescription())){
-           errors.add("The required field `description` isn't set on the incoming file.");
-       }
-
-       if(isNotSet(appDescriptor.getIconUrl())){
-           errors.add("The required field `iconUrl` isn't set on the incoming file.");
-       }
-
-       if(!isSet(appDescriptor.getAllowExtraParameters())){
-           errors.add("The required boolean field `allowExtraParameters` isn't set on the incoming file.");
-       }
-
-       if(!isSet(appDescriptor.getParams())){
-           errors.add("The required field `params` isn't set on the incoming file.");
-       }
-
-       for (final Map.Entry<String, ParamDescriptor> entry : appDescriptor.getParams().entrySet()) {
-           errors.addAll(validateParamDescriptor(entry.getKey(), entry.getValue()));
-       }
-
-       if(!errors.isEmpty()){
-           throw new DotDataValidationException(String.join(" \n", errors));
-       }
-
-       return true;
-
-   }
-
-    /**
-     * internal param validator
-     * @param name
-     * @param descriptor
-     * @return
-     */
-    private List<String> validateParamDescriptor(final String name,
-            final ParamDescriptor descriptor) {
-
-        final List<String> errors = new LinkedList<>();
-
-        if (isNotSet(name)) {
-            errors.add("Param descriptor is missing required  field `name` .");
-        }
-
-        if (DESCRIPTOR_NAME_MAX_LENGTH < name.length()) {
-            errors.add(String.format("`%s`: exceeds %d chars length.", name,
-                    DESCRIPTOR_NAME_MAX_LENGTH));
-        }
-
-        if (null == descriptor.getValue()) {
-            errors.add(String.format(
-                    "`%s`: is missing required field `value` or a value hasn't been set. Value is mandatory. ",
-                    name));
-        }
-
-        if (isNotSet(descriptor.getHint())) {
-            errors.add(String.format("Param `%s`: is missing required field `hint` .", name));
-        }
-
-        if (isNotSet(descriptor.getLabel())) {
-            errors.add(String.format("Param `%s`: is missing required field `hint` .", name));
-        }
-
-        if (null == descriptor.getType()) {
-            errors.add(String.format(
-                    "Param `%s`: is missing required field `type` (STRING|BOOL|SELECT) .",
-                    name));
-        }
-
-        if (!isSet(descriptor.getRequired())) {
-            errors.add(
-                    String.format("Param `%s`: is missing required field `required` (true|false) .",
-                            name));
-        }
-
-        if (!isSet(descriptor.getHidden())) {
-            errors.add(
-                    String.format("Param `%s`: is missing required field `hidden` (true|false) .",
-                            name));
-        }
-
-        if (isSet(descriptor.getValue()) && StringPool.NULL
-                .equalsIgnoreCase(descriptor.getValue().toString()) && descriptor.isRequired()) {
-            errors.add(String.format(
-                    "Null isn't allowed as the default value on required params see `%s`. ",
-                    name)
-            );
-        }
-
-        if (Type.BOOL.equals(descriptor.getType())) {
-            if (isSet(descriptor.getHidden())
-                    && descriptor.isHidden()) {
-                errors.add(String.format(
-                        "Param `%s`: Bool params can not be marked hidden. The combination (Bool + Hidden) isn't allowed.",
-                        name));
-            }
-
-            if (isSet(descriptor.getValue())
-                    && !isBoolString(descriptor.getValue().toString())) {
-                errors.add(String.format(
-                        "Boolean Param `%s` has a default value `%s` that can not be parsed to bool (true|false).",
-                        name, descriptor.getValue()));
-            }
-        }
-
-        if(Type.STRING.equals(descriptor.getType()) && !(descriptor.getValue() instanceof String)){
-                errors.add(String.format(
-                        "Value Param `%s` has a default value `%s` that isn't a string .",
-                        name, descriptor.getValue()));
-        }
-
-        if (Type.SELECT.equals(descriptor.getType())) {
-
-            if (isSet(descriptor.getHidden()) && descriptor.isHidden()) {
-                errors.add(String.format(
-                        "Param `%s`: List params can not be marked hidden. The combination (List + Hidden) isn't allowed.",
-                        name));
-            }
-
-            if (!(descriptor.getValue() instanceof List)) {
-                errors.add(String.format(
-                        " As param `%s`:  is marked as `List` the field value is expected to hold a list of objects. ",
-                        name));
-            } else {
-                final int minSelectedElements = 1;
-                int selectedCount = 0;
-                final List list = (List) descriptor.getValue();
-                for (final Object object : list) {
-                    if (!(object instanceof Map)) {
-                        errors.add(String.format(
-                                "Malformed list. Param: `%s` is marked as `List` therefore field `value` is expected to have a list of objects. ",
-                                name));
-                    } else {
-                        final Map map = (Map) object;
-                        if (!map.containsKey("label") || !map.containsKey("value") ) {
-                            errors.add(String.format("Malformed list. Param: `%s`. Every entry of the `List` has to have the following fields (`label`,`value`). ", name));
-                        }
-                         if(map.containsKey("selected")){
-                             selectedCount++;
-                         }
-                    }
-                }
-                if(selectedCount > minSelectedElements ){
-                    errors.add(String.format("Malformed list. Param: `%s`. There must be only 1 item marked as selected ", name));
-                }
-            }
-        }
-
-        return errors;
-    }
-
-    /**
-     * Verifies if a string can be parsed to boolean safely.
-     * @param value
-     * @return
-     */
-    private boolean isBoolString(final String value){
-      return Boolean.TRUE.toString().equalsIgnoreCase(value) || Boolean.FALSE.toString().equalsIgnoreCase(value);
-   }
-
-    /**
      * Method meant to to be consumed from a delete site event.
      * @param host
      * @param user
@@ -1028,90 +676,6 @@ public class AppsAPIImpl implements AppsAPI {
        secretsStore.backupAndRemoveKeyStore();
        //Clear cache forces reloading the yml app descriptors.
        appsCache.clearCache();
-    }
-
-    /**
-     * {@inheritDoc}
-     * @param params
-     * @param appDescriptor
-     */
-    public void validateForSave(final Map<String, Optional<char[]>> params,
-            final AppDescriptor appDescriptor) {
-
-        //Param/Property names are case sensitive.
-        final Map<String, ParamDescriptor> appDescriptorParams = appDescriptor.getParams();
-
-        for (final Entry<String, ParamDescriptor> descriptorParam : appDescriptorParams.entrySet()) {
-            final String describedParamName = descriptorParam.getKey();
-            //initialize to null so it is not found in the params map it means it wasn't sent.
-            char[] input = null;
-            if (params.containsKey(describedParamName)) {
-            // if the key is found then verify if there's an actual value or else null
-                final Optional<char[]> optionalChars = params.get(describedParamName);
-                input = optionalChars.orElse(null);
-            }
-            if (descriptorParam.getValue().isRequired() && (input == null || isNotSet(input))) {
-                throw new IllegalArgumentException(
-                        String.format(
-                                "Param `%s` is marked required in the descriptor but does not come with a value.",
-                                describedParamName
-                        )
-                );
-            }
-
-            if (null == input) {
-                //Param wasn't sent but it doesn't matter since it isn't required.
-                Logger.debug(AppsAPIImpl.class, () -> String
-                        .format("Non required param `%s` was set.",
-                                describedParamName));
-                continue;
-            }
-
-            if (Type.BOOL.equals(descriptorParam.getValue().getType()) && UtilMethods
-                    .isSet(input)) {
-                final String asString = new String(input);
-                final boolean bool = (asString.equalsIgnoreCase(Boolean.TRUE.toString())
-                        || asString.equalsIgnoreCase(Boolean.FALSE.toString()));
-                if (!bool) {
-                    throw new IllegalArgumentException(
-                            String.format(
-                                    "Can not convert value `%s` to type BOOL for param `%s`.",
-                                    asString, describedParamName
-                            )
-                    );
-                }
-            }
-
-            if (Type.SELECT.equals(descriptorParam.getValue().getType()) && UtilMethods
-                    .isSet(input)) {
-                final List<Map> list = descriptorParam.getValue().getList();
-                final Set<String> values = list.stream().filter(map -> null != map.get("value"))
-                        .map(map -> map.get("value").toString()).collect(Collectors.toSet());
-                final String asString = new String(input);
-                if (!values.contains(asString)) {
-                    throw new IllegalArgumentException(
-                            String.format(
-                                    "Can not find value `%s` in the list of permitted values `%s`.",
-                                    asString, describedParamName
-                            )
-                    );
-                }
-            }
-        }
-
-        if (!appDescriptor.isAllowExtraParameters()) {
-            final SetView<String> extraParamsFound = Sets
-                    .difference(params.keySet(), appDescriptorParams.keySet());
-
-            if (!extraParamsFound.isEmpty()) {
-                throw new IllegalArgumentException(
-                        String.format(
-                                "Unknown additional params `%s` not allowed by the app descriptor.",
-                                String.join(", ", extraParamsFound)
-                        )
-                );
-            }
-        }
     }
 
 
@@ -1149,26 +713,9 @@ public class AppsAPIImpl implements AppsAPI {
 
         Logger.info(AppsAPIImpl.class," exporting : "+exportedSecrets);
 
-        final File tempFile = File.createTempFile("secretsExport", ".tmp");
-        try {
-            writeObject(exportedSecrets, tempFile.toPath());
-            final byte[] bytes = Files.readAllBytes(tempFile.toPath());
-            try {
-                final File file = File.createTempFile("secrets", ".export");
-                file.deleteOnExit();
-                final byte[] encrypted = AppsUtil.encrypt(key, bytes);
-                final Path path = file.toPath();
-                try (OutputStream outputStream = Files.newOutputStream(path)) {
-                    outputStream.write(encrypted);
-                    return path;
-                }
-            } catch (EncryptorException e) {
-                throw new DotDataException(e);
-            }
-        } finally {
-            tempFile.delete();
-        }
+        return exportSecret(exportedSecrets, key);
     }
+
 
     /**
      * constructs the Import export object
@@ -1215,63 +762,7 @@ public class AppsAPIImpl implements AppsAPI {
         return new AppsSecretsImportExport(exportedSecrets);
     }
 
-    /**
-     * Takes a wrapping object that encapsulates all entries an write'em out ino a stream
-     * @param bean
-     * @param file
-     * @throws IOException
-     */
-    private void writeObject(final AppsSecretsImportExport bean, final Path file)
-            throws IOException {
-        try (OutputStream outputStream = Files.newOutputStream(file)) {
-            try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(outputStream)) {
-                objectOutputStream.writeObject(bean);
-            }
-        }
-    }
 
-    /**
-     * {@inheritDoc}
-     * @param incomingFile
-     * @param key
-     * @param user
-     * @return
-     * @throws DotDataException
-     * @throws DotSecurityException
-     * @throws IOException
-     * @throws EncryptorException
-     */
-
-    Map<String, List<AppSecrets>> importSecrets(final Path incomingFile, final Key key,
-            final User user)
-            throws DotDataException, DotSecurityException, IOException {
-        if(!user.isAdmin()){
-            throw new DotSecurityException("Only Admins are allowed to perform an export operation.");
-        }
-
-        if(!licenseValiditySupplier.hasValidLicense()){
-           throw new InvalidLicenseException("Apps requires of an enterprise level license.");
-        }
-
-        final byte[] encryptedBytes = Files.readAllBytes(incomingFile);
-        final byte[] decryptedBytes;
-        try {
-            decryptedBytes = AppsUtil.decrypt(key, encryptedBytes);
-        }catch (EncryptorException e){
-            throw new IllegalArgumentException("An error occurred while decrypting file contents. ",e.getCause());
-        }
-        final File importFile = File.createTempFile("secrets", "export");
-        try (OutputStream outputStream = Files.newOutputStream(importFile.toPath())) {
-            outputStream.write(decryptedBytes);
-        }
-       final AppsSecretsImportExport importExport;
-       try {
-           importExport = readObject(importFile.toPath());
-           return importExport.getSecrets();
-       } catch (ClassNotFoundException e) {
-           throw new DotDataException(e);
-       }
-    }
 
     /**
      * {@inheritDoc}
@@ -1282,16 +773,22 @@ public class AppsAPIImpl implements AppsAPI {
      * @throws DotDataException
      * @throws DotSecurityException
      * @throws IOException
-     * @throws EncryptorException
      */
     public int importSecretsAndSave(final Path incomingFile, final Key key, final User user)
-            throws DotDataException, DotSecurityException, IOException, EncryptorException {
-        final String failSilentlyMessage = "These exceptions can be ignored by setting the property `APPS_IMPORT_FAIL_SILENTLY` to true.";
+            throws DotDataException, DotSecurityException, IOException {
 
+        if(!user.isAdmin()){
+            throw new DotSecurityException("Only Admins are allowed to perform an export operation.");
+        }
+
+        if(!licenseValiditySupplier.hasValidLicense()){
+            throw new InvalidLicenseException("Apps requires of an enterprise level license.");
+        }
+
+        final String failSilentlyMessage = "These exceptions can be ignored by setting the property `APPS_IMPORT_FAIL_SILENTLY` to true.";
         final boolean failSilently = Config.getBooleanProperty(APPS_IMPORT_FAIL_SILENTLY, false);
         int count = 0;
-        final Map<String, List<AppSecrets>> importedSecretsBySiteId = importSecrets(incomingFile,
-                key, user);
+        final Map<String, List<AppSecrets>> importedSecretsBySiteId = importSecrets(incomingFile, key);
         Logger.info(AppsAPIImpl.class,
                 "Number of secrets found: " + importedSecretsBySiteId.size());
         for (final Entry<String, List<AppSecrets>> importEntry : importedSecretsBySiteId
@@ -1359,36 +856,6 @@ public class AppsAPIImpl implements AppsAPI {
           appsCache.flushSecret();
         }
         return count;
-    }
-
-    /**
-     * Map of optionals. Common portable format
-     */
-    private Map<String, Optional<char[]>> mapForValidation(final AppSecrets appSecrets) {
-        return appSecrets.getSecrets().entrySet().stream()
-            .collect(Collectors
-                .toMap(Entry::getKey,
-                secretEntry -> {
-                    final Secret value = secretEntry.getValue();
-                    return value == null ? Optional.empty()
-                            : Optional.of(value.getValue());
-                })
-            );
-    }
-
-    /**
-     * Reads the exported file stream
-     * and returns a wrapper that contains all entries.
-     * @param importFile
-     * @return
-     * @throws IOException
-     * @throws ClassNotFoundException
-     */
-    private AppsSecretsImportExport readObject(final Path importFile)
-            throws IOException, ClassNotFoundException {
-        try(InputStream inputStream = Files.newInputStream(importFile)){
-            return (AppsSecretsImportExport)new ObjectInputStream(inputStream).readObject();
-        }
     }
 
 }
