+++ dotCMS/src/main/java/com/dotcms/security/apps/AppsUtil.java
@@ -1,25 +1,45 @@
 package com.dotcms.security.apps;
 
+import static com.dotcms.security.apps.AppsAPI.DOT_GLOBAL_SERVICE;
+import static com.dotcms.security.apps.AppsAPI.HOST_SECRET_KEY_SEPARATOR;
 import static com.dotmarketing.util.UtilMethods.isNotSet;
+import static com.dotmarketing.util.UtilMethods.isSet;
 
 import com.dotcms.rest.api.v1.DotObjectMapperProvider;
+import com.dotmarketing.beans.Host;
+import com.dotmarketing.business.APILocator;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.util.Config;
 import com.dotmarketing.util.Logger;
+import com.dotmarketing.util.UtilMethods;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Sets;
+import com.google.common.collect.Sets.SetView;
 import com.google.common.io.ByteStreams;
 import com.liferay.util.Base64;
 import com.liferay.util.Encryptor;
 import com.liferay.util.EncryptorException;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
+import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.security.Key;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Optional;
+import java.util.Set;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 import javax.crypto.Cipher;
 import javax.crypto.spec.SecretKeySpec;
 import org.apache.commons.io.input.CharSequenceInputStream;
@@ -30,6 +50,32 @@ public class AppsUtil {
     private static final ObjectMapper mapper = DotObjectMapperProvider.getInstance().getDefaultObjectMapper();
     private static final String APPS_IMPORT_EXPORT_DEFAULT_PASSWORD = "APPS_IMPORT_EXPORT_DEFAULT_PASSWORD";
 
+
+    /**
+     * One single method takes care of building the internal-key
+     */
+    public static String internalKey(final String serviceKey, final Host host) {
+        return internalKey(serviceKey, host == null ? null : host.getIdentifier());
+    }
+
+    /**
+     * Given a service key and an identifier this builds an internal key composed by the two values concatenated
+     * And lowercased.
+     * Like `5e096068-edce-4a7d-afb1-95f30a4fa80e:serviceKeyNameXYZ`
+     * @param serviceKey
+     * @param hostIdentifier
+     * @return
+     */
+    public static String internalKey(final String serviceKey, final String hostIdentifier) {
+        // if Empty ServiceKey is passed everything will be set under systemHostIdentifier:dotCMSGlobalService
+        //Otherwise the internal Key will look like:
+        // `5e096068-edce-4a7d-afb1-95f30a4fa80e:serviceKeyNameXYZ` where the first portion is the hostId
+        final String key = isSet(serviceKey) ? serviceKey : DOT_GLOBAL_SERVICE;
+        final String identifier =
+                (null == hostIdentifier) ? APILocator.systemHost().getIdentifier() : hostIdentifier;
+        return (identifier + HOST_SECRET_KEY_SEPARATOR + key).toLowerCase();
+    }
+
     /**
      * Given the AppSecrets this will return a char array representing the deserialized json object.
      * No strings are created in the transformation process.
@@ -37,7 +83,7 @@ public class AppsUtil {
      * @return
      * @throws DotDataException
      */
-    static char[] toJsonAsChars(final AppSecrets object) throws DotDataException {
+    public static char[] toJsonAsChars(final AppSecrets object) throws DotDataException {
         try {
             final byte [] bytes = mapper.writeValueAsBytes(object);
             return bytesToCharArrayUTF(bytes);
@@ -240,6 +286,205 @@ public class AppsUtil {
             password = Config.getStringProperty(APPS_IMPORT_EXPORT_DEFAULT_PASSWORD);
         }
         return password;
+    }
+
+    /**
+     * Ths will return a list of AppSecrets arranged in a map whose key is the site-id
+     * Therefore we have a Secrets by site
+     * @param incomingFile
+     * @param key
+     * @return
+     * @throws DotDataException
+     * @throws IOException
+     * @throws EncryptorException
+     */
+
+     public static Map<String, List<AppSecrets>> importSecrets(final Path incomingFile, final Key key)
+            throws DotDataException, IOException {
+
+        final byte[] encryptedBytes = Files.readAllBytes(incomingFile);
+        final byte[] decryptedBytes;
+        try {
+            decryptedBytes = AppsUtil.decrypt(key, encryptedBytes);
+        }catch (EncryptorException e){
+            throw new IllegalArgumentException("An error occurred while decrypting file contents. ",e.getCause());
+        }
+        final File importFile = File.createTempFile("secrets", "export");
+        try (OutputStream outputStream = Files.newOutputStream(importFile.toPath())) {
+            outputStream.write(decryptedBytes);
+        }
+        final AppsSecretsImportExport importExport;
+        try {
+            importExport = readObject(importFile.toPath());
+            return importExport.getSecrets();
+        } catch (ClassNotFoundException e) {
+            throw new DotDataException(e);
+        }
+    }
+
+    /**
+     * Reads the exported file stream
+     * and returns a wrapper that contains all entries.
+     * @param importFile
+     * @return
+     * @throws IOException
+     * @throws ClassNotFoundException
+     */
+    private static AppsSecretsImportExport readObject(final Path importFile)
+            throws IOException, ClassNotFoundException {
+        try(InputStream inputStream = Files.newInputStream(importFile)){
+            return (AppsSecretsImportExport)new ObjectInputStream(inputStream).readObject();
+        }
+    }
+
+    /**
+     *
+     * @param exportedSecrets
+     * @param key
+     * @return
+     * @throws IOException
+     * @throws DotDataException
+     */
+    static Path exportSecret(final AppsSecretsImportExport exportedSecrets,final Key key)
+            throws IOException, DotDataException {
+        final File tempFile = File.createTempFile("secretsExport", ".tmp");
+        try {
+            writeObject(exportedSecrets, tempFile.toPath());
+            final byte[] bytes = Files.readAllBytes(tempFile.toPath());
+            try {
+                final File file = File.createTempFile("secrets", ".export");
+                file.deleteOnExit();
+                final byte[] encrypted = AppsUtil.encrypt(key, bytes);
+                final Path path = file.toPath();
+                try (OutputStream outputStream = Files.newOutputStream(path)) {
+                    outputStream.write(encrypted);
+                    return path;
+                }
+            } catch (EncryptorException e) {
+                throw new DotDataException(e);
+            }
+        } finally {
+            tempFile.delete();
+        }
+    }
+
+    /**
+     * Takes a wrapping object that encapsulates all entries an write'em out ino a stream
+     * @param bean
+     * @param file
+     * @throws IOException
+     */
+    private static void writeObject(final AppsSecretsImportExport bean, final Path file)
+            throws IOException {
+        try (OutputStream outputStream = Files.newOutputStream(file)) {
+            try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(outputStream)) {
+                objectOutputStream.writeObject(bean);
+            }
+        }
+    }
+
+    /**
+     * Map of optionals. Common portable format
+     */
+    public static  Map<String, Optional<char[]>> mapForValidation(final AppSecrets appSecrets) {
+        return appSecrets.getSecrets().entrySet().stream()
+                .collect(Collectors
+                        .toMap(Entry::getKey,
+                                secretEntry -> {
+                                    final Secret value = secretEntry.getValue();
+                                    return value == null ? Optional.empty()
+                                            : Optional.of(value.getValue());
+                                })
+                );
+    }
+
+    /**
+     * Validate the incoming params and match them with the params described by the respective appDescriptor yml.
+     * This method takes a Map of Optional<char[]> As this is a middle ground object representation
+     * that can be mapped from a saved  AppSecrets or an incoming SecretForm
+     * if the param isn't included in the map it means it wasn't sent..
+     * if the param was sent empty that would be represented as an empty optional.
+     * I'm using optional since null vales on map triggers warnings
+     * @param params
+     * @param appDescriptor
+     */
+    public static void validateForSave(final Map<String, Optional<char[]>> params,
+            final AppDescriptor appDescriptor) {
+
+        //Param/Property names are case sensitive.
+        final Map<String, ParamDescriptor> appDescriptorParams = appDescriptor.getParams();
+
+        for (final Entry<String, ParamDescriptor> descriptorParam : appDescriptorParams.entrySet()) {
+            final String describedParamName = descriptorParam.getKey();
+            //initialize to null so it is not found in the params map it means it wasn't sent.
+            char[] input = null;
+            if (params.containsKey(describedParamName)) {
+                // if the key is found then verify if there's an actual value or else null
+                final Optional<char[]> optionalChars = params.get(describedParamName);
+                input = optionalChars.orElse(null);
+            }
+            if (descriptorParam.getValue().isRequired() && (input == null || isNotSet(input))) {
+                throw new IllegalArgumentException(
+                        String.format(
+                                "Param `%s` is marked required in the descriptor but does not come with a value.",
+                                describedParamName
+                        )
+                );
+            }
+
+            if (null == input) {
+                //Param wasn't sent but it doesn't matter since it isn't required.
+                Logger.debug(AppsAPIImpl.class, () -> String
+                        .format("Non required param `%s` was set.",
+                                describedParamName));
+                continue;
+            }
+
+            if (Type.BOOL.equals(descriptorParam.getValue().getType()) && UtilMethods
+                    .isSet(input)) {
+                final String asString = new String(input);
+                final boolean bool = (asString.equalsIgnoreCase(Boolean.TRUE.toString())
+                        || asString.equalsIgnoreCase(Boolean.FALSE.toString()));
+                if (!bool) {
+                    throw new IllegalArgumentException(
+                            String.format(
+                                    "Can not convert value `%s` to type BOOL for param `%s`.",
+                                    asString, describedParamName
+                            )
+                    );
+                }
+            }
+
+            if (Type.SELECT.equals(descriptorParam.getValue().getType()) && UtilMethods
+                    .isSet(input)) {
+                final List<Map> list = descriptorParam.getValue().getList();
+                final Set<String> values = list.stream().filter(map -> null != map.get("value"))
+                        .map(map -> map.get("value").toString()).collect(Collectors.toSet());
+                final String asString = new String(input);
+                if (!values.contains(asString)) {
+                    throw new IllegalArgumentException(
+                            String.format(
+                                    "Can not find value `%s` in the list of permitted values `%s`.",
+                                    asString, describedParamName
+                            )
+                    );
+                }
+            }
+        }
+
+        if (!appDescriptor.isAllowExtraParameters()) {
+            final SetView<String> extraParamsFound = Sets
+                    .difference(params.keySet(), appDescriptorParams.keySet());
+
+            if (!extraParamsFound.isEmpty()) {
+                throw new IllegalArgumentException(
+                        String.format(
+                                "Unknown additional params `%s` not allowed by the app descriptor.",
+                                String.join(", ", extraParamsFound)
+                        )
+                );
+            }
+        }
     }
 
 }
